+++ images/the-web-cloud-server/src/Logger.ts
@@ -7,20 +7,20 @@ export class Logger {
         this.channel = channel;
     }
 
-    public info(msg: any): void {
+    public info(msg: Error | string): void {
         process.stdout.write(this.stringifyMessage(msg));
     }
 
-    public error(msg: any): void {
+    public error(msg: Error | string): void {
         process.stderr.write(this.stringifyMessage(msg));
     }
 
-    public critical(msg: any): void {
+    public critical(msg: Error | string): void {
         this.error(msg);
         writeFileSync("/dev/termination-log", this.stringifyMessage(msg));
     }
 
-    private stringifyMessage(msg: any): string {
+    private stringifyMessage(msg: Error | string): string {
         return `[${this.channel}] ${msg.toString()}\n`;
     }
 }