+++ images/the-web-cloud-tester/src/HomepagePayload.ts
@@ -1,0 +1,5 @@
+export interface HomepagePayload {
+    message: string;
+    method: string;
+    url: string;
+}