+++ images/the-web-cloud-tester/src/IHomepagePayload.ts
@@ -1,5 +1,0 @@
-export interface IHomepagePayload {
-    message: string
-    method: string
-    url: string
-}