+++ images/the-web-cloud-tester/tests/target-smoke-tests/homepage.test.ts
@@ -1,11 +1,11 @@
 import "jest";
 import fetch, { Response } from "node-fetch";
 import { config } from "../../src/config";
-import { IHomepagePayload } from "../../src/IHomepagePayload";
+import { HomepagePayload } from "../../src/HomepagePayload";
 
 test("should return welcome message", async (): Promise<void> => {
     const res: Response = await fetch(config.targetEndpoint);
-    const payload: IHomepagePayload = await res.json();
+    const payload: HomepagePayload = await res.json();
 
     expect(res.status).toBe(200);
     expect(res.headers.has("content-type")).toBe(true);