+++ charts/the-web-cloud-platform/pipeline/run-system-tests.sh
@@ -4,11 +4,11 @@ set -e
 testerPodName="$(kubectl \
   --namespace="default" \
   get pod \
+  --selector="app.kubernetes.io/instance=the-web-cloud-platform-0,app.kubernetes.io/name=the-web-cloud-app,app.kubernetes.io/component=tester" \
   --output="jsonpath={.items[0].metadata.name}" \
-  --selector="app.kubernetes.io/instance=the-web-cloud-platform-0,app.kubernetes.io/name=the-web-cloud-tester,app.kubernetes.io/component=tester" \
 )"
 
 kubectl \
   --namespace="default" \
   exec "$testerPodName" -- \
-  npm run target-api-tests
+  npm run api-tests