+++ images/the-web-cloud-app/src/MessagePayload.ts
@@ -1,0 +1,5 @@
+export interface MessagePayload {
+    method: string;
+    url: string;
+    message: string;
+}