+++ images/the-web-cloud-tester/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  testEnvironment: "node",
-  roots: [
-    "<rootDir>/tests",
-  ],
-  preset: "ts-jest",
-};