+++ images/the-web-cloud-tester/pipeline.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env sh
-set -e
-
-./pipeline/build.sh
-./pipeline/.reset.sh
-./pipeline/run-vulnerabilities-scanning.sh
-./pipeline/lint.sh
-./pipeline/run-unit-tests.sh
-./pipeline/publish.sh