+++ images/the-web-cloud-tester/pipeline/run-unit-tests.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env sh
-set -e
-
-docker-compose run --rm --entrypoint='' dist bash -ce "
-  npm run unit-tests
-"