+++ images/the-web-cloud-tester/tests/integration-tests/dummy.test.ts
@@ -1,5 +1,0 @@
-import "jest";
-
-test("test nothing", (): void => {
-    expect(true).toBe(true);
-});