+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Use `node-fetch` instead of `Axios`.
 
+### Fixed
+
+- Use recommended `@typescript-eslint` rules.
+
 ## [0.15.2] - 2019-01-26
 
 ### Changed